updated sounds and tome particle
music now doesnt reset at a new scene (need to come back on it to make it stop on last scene), sound on trigger enter for evil tome and door. on level need to click on tome to change scene but it's a UI button so it needs to be changed later.